Move IMDBCard to "components"
